Moved mock application to new Test directory
 - Created new test folder for functional tests;
 - Moved mock triangle implementation to 'helloTriangle' func test;
 - Modified CMake accordingly (added TEST_ENABLED config flag);
 - Rendering structures in Application made unique via static keyword;
 - Tests are accessed through Body, for now;
