--- skelly/examples/sanity/sanity.cpp
@@ -1,10 +1,0 @@
-#include "sanity.h"
-#include <stdio.h>
-
-
-namespace RESkelly {
-
-    void sanity () {
-        printf("Skelly sanity message reached. All set! :)\n");
-    }
-}
