--- skelly/examples/sanity/sanity.h
@@ -1,5 +1,0 @@
-#pragma once
-
-namespace RESkelly {
-    void sanity(); 
-}
