+++ skelly/test/helloTriangle/helloTriangle.h
@@ -1,0 +1,10 @@
+#include <application.h>
+
+namespace test {
+
+    class HelloTriangle {
+        public:
+            void helloTriangle(skelly::Application& app);
+    };
+
+}
