+++ skelly/test/skellyExamples.h
@@ -1,0 +1,1 @@
+#include <helloTriangle/helloTriangle.h>
