+++ skelly/utils/application/application.cpp
@@ -25,9 +25,12 @@
 
 namespace skelly {
     Application* Application::_s_instance;
+    // DN: Since the App is supposed to be unique, are single static shared pointers for
+    // vertexArray and shader enough in the long term?
+    std::shared_ptr<VertexArray> Application::_m_vertexArray;
+    std::shared_ptr<Shader> Application::_m_shader;
 
     // Lifetime handlers
-
     // If you modify the constructor behavior, be careful! You need to create the window 
     // context before creating the layers and overlays.
     Application::Application() {
@@ -49,65 +52,7 @@ namespace skelly {
         popOverlay(_m_imguiLayer.get());
     }
     
-    void Application::mockTriangle() {
-        // triangle drawing example
-
-        _m_vertexArray.reset(VertexArray::create());
-
-        float vertices[6*7] = {
-            -0.5f, -0.5f, 0.0f, 0.4f, 0.1f, 0.6f, 1.0f,
-            0.5f,  -0.5f, 0.0f, 1.0f, 1.0f, 0.4f, 1.0f,
-            0.0f,  0.8f,  0.0f, 0.4f, 0.1f, 0.6f, 1.0f,
-        };
-
-        std::shared_ptr<VertexBuffer> vertexBuffer;
-        vertexBuffer.reset(VertexBuffer::create(vertices, sizeof(vertices)));
-
-        BufferLayout layout = {
-            { ShaderDataType::Float3, "a_Position" },
-            { ShaderDataType::Float4, "a_Color" },
-        };
-        vertexBuffer->setLayout(layout);
-
-        _m_vertexArray->addVertexBuffer(vertexBuffer);
-
-        // index buffer creation
-        uint32_t indices[3] = {0, 1, 2};
-        std::shared_ptr<IndexBuffer> indexBuffer;
-        indexBuffer.reset(IndexBuffer::create(indices, sizeof(indices) / sizeof(uint32_t)));
-        _m_vertexArray->addIndexBuffer(indexBuffer);
-
-        std::string vertexSrc = R"(
-            #version 450 core
-
-            layout(location = 0) in vec3 a_Position;
-            layout(location = 1) in vec4 a_Color;
-
-            out vec4 v_Color;
-
-            void main() {
-                v_Color = a_Color;
-                gl_Position = vec4(a_Position, 1.0);
-            }
-        )";
-
-        std::string fragmentSrc = R"(
-            #version 450 core
-
-            layout(location = 0) out vec4 color;
-
-            in vec4 v_Color;
-
-            void main() {
-                color = v_Color;
-            }
-        )";
-
-        _m_shader.reset(Shader::create(vertexSrc, fragmentSrc));
-    }
-
     // layer stack handlers
-
     void Application::pushLayer(Layer* layer) {
         SKELLY_LOG_TRACE("Creating new layer: {0}", layer->getName());
         _m_layerStack.pushLayer(layer);
@@ -130,7 +75,6 @@ namespace skelly {
     }
 
     // event handlers
-
     void Application::onEvent(Event& e) {
         EventDispatcher dispatcher(e);
         dispatcher.dispatch<WindowCloseEvent>(BIND_EVENT_FN(Application::_m_onWindowClose));
@@ -147,7 +91,6 @@ namespace skelly {
     }    
 
     // main application loop
-
     void Application::run() {
         while(_m_running) {
